+++ engine-tests/src/utils/workspace.rs
@@ -5,8 +5,12 @@ use crate::utils::solidity::erc20::{ERC20Constructor, ERC20};
 /// it does not execute promises; but `aurora-workspaces` does.
 use crate::utils::AuroraRunner;
 use aurora_engine_types::account_id::AccountId;
+use aurora_engine_types::parameters::connector::Erc20Metadata;
 #[cfg(feature = "ext-connector")]
 use aurora_engine_types::parameters::connector::{FungibleTokenMetadata, WithdrawSerializeType};
+use aurora_engine_types::parameters::engine::{
+    DeployErc20TokenArgs, DeployErc20TokenArgsWithMetadata,
+};
 use aurora_engine_types::types::Address;
 use aurora_engine_types::U256;
 use aurora_engine_workspace::account::Account;
@@ -124,14 +128,25 @@ pub async fn create_sub_account(
 pub async fn deploy_erc20_from_nep_141(
     nep_141_account: &str,
     aurora: &EngineContract,
+    erc20_metadata: Option<Erc20Metadata>,
 ) -> anyhow::Result<ERC20> {
-    let nep141_account_id = nep_141_account.parse().unwrap();
-    let result = aurora
-        .deploy_erc20_token(nep141_account_id)
-        .max_gas()
-        .transact()
-        .await
-        .unwrap();
+    let nep141 = nep_141_account.parse().unwrap();
+    let result = if let Some(metadata) = erc20_metadata {
+        aurora
+            .deploy_erc20_token(DeployErc20TokenArgs::WithMetadata(
+                DeployErc20TokenArgsWithMetadata { nep141, metadata },
+            ))
+            .max_gas()
+            .transact()
+            .await?
+    } else {
+        aurora
+            .deploy_erc20_token_legacy(nep141)
+            .max_gas()
+            .transact()
+            .await?
+    };
+
     assert!(result.is_success());
     let address = result.into_value();
     let abi = ERC20Constructor::load().0.abi;