+++ CHANGES.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Additions
+
+- Added a possibility to deploy ERC-20 token with metadata by [@aleksuss]. ([#1008])
+
 ### Changes
 
 - Reworked the logic related to the silo mode by [@aleksuss]. ([#1005])
 
 [#1005]: https://github.com/aurora-is-near/aurora-engine/pull/1005
+[#1008]: https://github.com/aurora-is-near/aurora-engine/pull/1008
 
 ## [3.8.0] 2025-02-05